move functions to allocator struct
