+++ src/vulkan/visualizer.rs
@@ -1,7 +1,7 @@
 #![allow(clippy::new_without_default)]
 
 use super::Allocator;
-use crate::allocator::resolve_backtrace;
+use crate::allocator::{fmt_bytes, resolve_backtrace};
 use crate::visualizer::ColorScheme;
 use log::error;
 
@@ -310,7 +310,7 @@ impl AllocatorVisualizer {
 
         let mut window = imgui::Window::new(format!(
             "Allocation Breakdown ({})###allocation_breakdown_window",
-            crate::visualizer::fmt_bytes(total_size_in_bytes)
+            fmt_bytes(total_size_in_bytes)
         ))
         .position([20.0f32, 80.0f32], imgui::Condition::FirstUseEver)
         .size([460.0f32, 420.0f32], imgui::Condition::FirstUseEver);
@@ -386,49 +386,9 @@ impl AllocatorVisualizer {
                     }
 
                     ui.table_next_column();
-                    ui.text(crate::visualizer::fmt_bytes(alloc.size));
+                    ui.text(fmt_bytes(alloc.size));
                 }
             }
         });
-    }
-
-    #[allow(clippy::print_stdout)]
-    pub fn debug_pring_breakdown(allocator: &Allocator) {
-        let mut allocation_report = vec![];
-
-        for memory_type in &allocator.memory_types {
-            for block in memory_type.memory_blocks.iter().flatten() {
-                allocation_report.extend_from_slice(&block.sub_allocator.report_allocations())
-            }
-        }
-
-        let total_size_in_bytes = allocation_report.iter().map(|report| report.size).sum();
-
-        let mut allocation_report = allocation_report.iter().enumerate().collect::<Vec<_>>();
-        allocation_report.sort_by_key(|(_, alloc)| std::cmp::Reverse(alloc.size));
-
-        const MAX_NUM_CHARACTERS: usize = 40;
-
-        println!("================================================================");
-        println!(
-            "ALLOCATION BREAKDOWN ({})",
-            crate::visualizer::fmt_bytes(total_size_in_bytes)
-        );
-
-        for (idx, alloc) in &allocation_report {
-            let mut cloned_name = alloc.name.clone();
-            cloned_name.truncate(MAX_NUM_CHARACTERS);
-
-            let num_spaces = MAX_NUM_CHARACTERS - cloned_name.len();
-            let aligning_spaces = " ".repeat(num_spaces);
-
-            println!(
-                "\t\t{}\t- {}{}\t- {}",
-                idx,
-                cloned_name,
-                aligning_spaces,
-                crate::visualizer::fmt_bytes(alloc.size)
-            );
-        }
     }
 }
